requirements.txt
plotly==5.11.0
numpy==1.23.4
pandas==1.5.1
streamlit==1.20.0
pip==21.3.1
matplotlib==3.6.2
wordcloud==1.8.2.2
Pillow==9.0.0
streamlit-aggrid==0.3.3
DateTime==4.7
requests==2.28.1
